remove all '/m' handlers
Fixes #199
