+++ controllers/invite/invite.go
@@ -19,7 +19,6 @@ package invite
 import (
 	"errors"
 	"fmt"
-	"html/template"
 	"net/http"
 	"strings"
 
@@ -28,16 +27,8 @@ import (
 
 	"github.com/santiaago/purple-wing/helpers"
 	"github.com/santiaago/purple-wing/helpers/log"
-
-	"github.com/santiaago/purple-wing/helpers/auth"
 	templateshlp "github.com/santiaago/purple-wing/helpers/templates"
 )
-
-type InviteForm struct {
-	Name       string
-	EmailsList string
-	Error      string
-}
 
 const inviteMessage = `
 Hi,
@@ -52,56 +43,6 @@ Your friends @ purple-wing
 
 
 `
-
-// Email invite handler
-func Email(w http.ResponseWriter, r *http.Request) {
-	c := appengine.NewContext(r)
-
-	var form InviteForm
-	form.Name = auth.CurrentUser(r, c).Name
-
-	if r.Method == "GET" {
-		form.EmailsList = ""
-		form.Error = ""
-	} else if r.Method == "POST" {
-		log.Infof(c, " Form Value = %v", r.FormValue("emails_area"))
-		form.EmailsList = r.FormValue("emails_area")
-
-		if len(form.EmailsList) <= 0 {
-			form.Error = "No email address has been entered"
-		} else {
-			emails := strings.Split(form.EmailsList, ",")
-
-			// validate emails
-			if !helpers.AreEmailsValid(emails) {
-				form.Error = "Your list of emails is not properly formatted"
-			} else {
-				url := fmt.Sprintf("http://%s/m/auth", r.Host)
-				for _, email := range emails {
-					msg := &mail.Message{
-						Sender:  "No Reply purple-wing <no-reply@purple-wing.com>",
-						To:      []string{email},
-						Subject: auth.CurrentUser(r, c).Name + " wants you to join Purple-wing!",
-						Body:    fmt.Sprintf(inviteMessage, url),
-					}
-
-					if err := mail.Send(c, msg); err != nil {
-						log.Errorf(c, " couldn't send email: %v", err)
-					}
-				}
-				http.Redirect(w, r, "/m/", http.StatusFound)
-				return
-			}
-		}
-	} else {
-		helpers.Error404(w)
-	}
-	t := template.Must(template.New("tmpl_invite_email").
-		ParseFiles("templates/invite/email.html"))
-
-	funcs := template.FuncMap{}
-	templateshlp.RenderWithData(w, r, c, t, form, funcs, "renderEmail")
-}
 
 // invite json handler
 func InviteJson(w http.ResponseWriter, r *http.Request) error {
