--- controllers/sessions/api_keys.go
@@ -1,21 +1,0 @@
-// Never add this file to source control
-
-package sessions
-
-const (
-	// GOOGLE
-	// Created at http://code.google.com/apis/console
-	// these identifies our app for the OAuth protocol.
-	GOOGLE_CLIENT_ID     = "36368233290.apps.googleusercontent.com"
-	GOOGLE_CLIENT_SECRET = "ZSU7Soyt3N2ipRddpghScXlx"
-
-	// TWITTER
-	// Created at https://dev.twitter.com
-	CONSUMER_KEY    = "bCXwcYTVzgONanLHAvg"
-	CONSUMER_SECRET = "qXXkPWOgCUjmTPPZ7Io7luF7rSLlFhe0NGVZ3QTRYI"
-
-	// FACEBOOK
-	// Created at https://developers.facebook.com/apps/
-	FACEBOOK_CLIENT_ID     = "232160743609875"
-	FACEBOOK_CLIENT_SECRET = "f440c8e62bb967c6ca5983cbe986b167"
-)
