+++ controllers/tournaments/tournaments.go
@@ -19,8 +19,6 @@ package tournaments
 import (
 	"encoding/json"
 	"errors"
-	"fmt"
-	"html/template"
 	"io/ioutil"
 	"net/http"
 	"time"
@@ -28,7 +26,6 @@ import (
 	"appengine"
 
 	"github.com/santiaago/purple-wing/helpers"
-	"github.com/santiaago/purple-wing/helpers/auth"
 	"github.com/santiaago/purple-wing/helpers/handlers"
 	"github.com/santiaago/purple-wing/helpers/log"
 	templateshlp "github.com/santiaago/purple-wing/helpers/templates"
@@ -43,234 +40,8 @@ import (
 	usermdl "github.com/santiaago/purple-wing/models/user"
 )
 
-type Form struct {
-	Name  string
-	Error string
-}
-
 type TournamentData struct {
 	Name string
-}
-
-type indexData struct {
-	Tournaments           []*tournamentmdl.Tournament
-	TournamentInputSearch string
-}
-
-// index tournaments handler
-func Index(w http.ResponseWriter, r *http.Request) {
-	c := appengine.NewContext(r)
-
-	var data indexData
-	if r.Method == "GET" {
-		tournaments := tournamentmdl.FindAll(c)
-
-		data.Tournaments = tournaments
-		data.TournamentInputSearch = ""
-
-	} else if r.Method == "POST" {
-		if query := r.FormValue("TournamentInputSearch"); len(query) == 0 {
-			http.Redirect(w, r, "tournaments", http.StatusFound)
-			return
-		} else {
-			words := helpers.SetOfStrings(query)
-			ids, err := tournamentinvmdl.GetIndexes(c, words)
-
-			if err != nil {
-				log.Errorf(c, " tournaments.Index, error occurred when getting indexes of words: %v", err)
-			}
-
-			result := searchmdl.TournamentScore(c, query, ids)
-
-			tournaments := tournamentmdl.ByIds(c, result)
-			data.Tournaments = tournaments
-			data.TournamentInputSearch = query
-		}
-	} else {
-		helpers.Error404(w)
-		return
-	}
-
-	t := template.Must(template.New("tmpl_tournament_index").
-		ParseFiles("templates/tournament/index.html"))
-
-	funcs := template.FuncMap{
-		"Tournaments": func() bool { return true },
-	}
-
-	templateshlp.RenderWithData(w, r, c, t, data, funcs, "renderTournamentIndex")
-}
-
-// new tournament handler
-func New(w http.ResponseWriter, r *http.Request) {
-	c := appengine.NewContext(r)
-
-	var form Form
-	if r.Method == "GET" {
-		form.Name = ""
-		form.Error = ""
-	} else if r.Method == "POST" {
-		form.Name = r.FormValue("Name")
-
-		if len(form.Name) <= 0 {
-			form.Error = "'Name' field cannot be empty"
-		} else if t := tournamentmdl.Find(c, "KeyName", helpers.TrimLower(form.Name)); t != nil {
-			form.Error = "That tournament name already exists."
-		} else {
-			tournament, err := tournamentmdl.Create(c, form.Name, "description foo", time.Now(), time.Now(), auth.CurrentUser(r, c).Id)
-			if err != nil {
-				log.Errorf(c, " error when trying to create a tournament: %v", err)
-			}
-			// redirect to the newly created tournament page
-			http.Redirect(w, r, "/m/tournaments/"+fmt.Sprintf("%d", tournament.Id), http.StatusFound)
-			return
-		}
-	} else {
-		helpers.Error404(w)
-		return
-	}
-
-	t := template.Must(template.New("tmpl_tournament_new").
-		ParseFiles("templates/tournament/new.html"))
-
-	funcs := template.FuncMap{}
-
-	templateshlp.RenderWithData(w, r, c, t, form, funcs, "renderTournamentNew")
-}
-
-// show tournament handler
-func Show(w http.ResponseWriter, r *http.Request) {
-	c := appengine.NewContext(r)
-
-	intID, err := handlers.PermalinkID(r, c, 3)
-	if err != nil {
-		log.Errorf(c, " Unable to find ID in request %v", r)
-		http.Redirect(w, r, "/m/tournaments/", http.StatusFound)
-		return
-	}
-
-	if r.Method == "GET" {
-		funcs := template.FuncMap{
-			"Joined":            func() bool { return tournamentmdl.Joined(c, intID, auth.CurrentUser(r, c).Id) },
-			"TeamJoined":        func(teamId int64) bool { return tournamentmdl.TeamJoined(c, intID, teamId) },
-			"IsTournamentAdmin": func() bool { return tournamentmdl.IsTournamentAdmin(c, intID, auth.CurrentUser(r, c).Id) },
-		}
-
-		t := template.Must(template.New("tmpl_tournament_show").
-			Funcs(funcs).
-			ParseFiles("templates/tournament/show.html",
-			"templates/tournament/participants.html",
-			"templates/tournament/teams.html",
-			"templates/tournament/candidateTeams.html"))
-
-		var tournament *tournamentmdl.Tournament
-		tournament, err = tournamentmdl.ById(c, intID)
-
-		if err != nil {
-			helpers.Error404(w)
-			return
-		}
-
-		participants := tournamentrelshlp.Participants(c, intID)
-		teams := tournamentrelshlp.Teams(c, intID)
-		candidateTeams := teammdl.Find(c, "AdminId", auth.CurrentUser(r, c).Id)
-
-		tournamentData := struct {
-			Tournament     *tournamentmdl.Tournament
-			Participants   []*usermdl.User
-			Teams          []*teammdl.Team
-			CandidateTeams []*teammdl.Team
-		}{
-			tournament,
-			participants,
-			teams,
-			candidateTeams,
-		}
-		templateshlp.RenderWithData(w, r, c, t, tournamentData, funcs, "renderTournamentShow")
-	}
-}
-
-//  Edit tournament handler
-func Edit(w http.ResponseWriter, r *http.Request) {
-	c := appengine.NewContext(r)
-
-	intID, err := handlers.PermalinkID(r, c, 3)
-	if err != nil {
-		http.Redirect(w, r, "/m/tournaments/", http.StatusFound)
-		return
-	}
-
-	if !tournamentmdl.IsTournamentAdmin(c, intID, auth.CurrentUser(r, c).Id) {
-		http.Redirect(w, r, "/m", http.StatusFound)
-		return
-	}
-
-	var tournament *tournamentmdl.Tournament
-	tournament, err = tournamentmdl.ById(c, intID)
-	if err != nil {
-		log.Errorf(c, " Tournament Edit handler: tournament not found. id: %v", intID)
-		helpers.Error404(w)
-		return
-	}
-
-	if r.Method == "GET" {
-
-		funcs := template.FuncMap{}
-
-		t := template.Must(template.New("tmpl_tournament_edit").
-			ParseFiles("templates/tournament/edit.html"))
-
-		templateshlp.RenderWithData(w, r, c, t, tournament, funcs, "renderTournamentEdit")
-		return
-	} else if r.Method == "POST" {
-
-		// only work on name other values should not be editable
-		editName := r.FormValue("Name")
-
-		if helpers.IsStringValid(editName) && editName != tournament.Name {
-			tournament.Name = editName
-			tournamentmdl.Update(c, intID, tournament)
-		} else {
-			log.Errorf(c, " cannot update %v", helpers.IsStringValid(editName))
-		}
-		url := fmt.Sprintf("/m/tournaments/%d", intID)
-		http.Redirect(w, r, url, http.StatusFound)
-		return
-	} else {
-		helpers.Error404(w)
-		return
-	}
-}
-
-// Tournament destroy handler
-func Destroy(w http.ResponseWriter, r *http.Request) {
-	c := appengine.NewContext(r)
-
-	tournamentID, err := handlers.PermalinkID(r, c, 4)
-	if err != nil {
-		http.Redirect(w, r, "/m/tournaments/", http.StatusFound)
-		return
-	}
-
-	if r.Method == "POST" {
-		// delete all tournament-user relationships
-		for _, participant := range tournamentrelshlp.Participants(c, tournamentID) {
-			if err := tournamentrelmdl.Destroy(c, tournamentID, participant.Id); err != nil {
-				log.Errorf(c, " error when trying to destroy tournament relationship: %v", err)
-			}
-		}
-		// delete all tournament-team relationships
-		for _, team := range tournamentrelshlp.Teams(c, tournamentID) {
-			if err := tournamentteamrelmdl.Destroy(c, tournamentID, team.Id); err != nil {
-				log.Errorf(c, " error when trying to destroy team relationship: %v", err)
-			}
-		}
-		// delete the tournament
-		tournamentmdl.Destroy(c, tournamentID)
-
-		http.Redirect(w, r, "/m/tournaments", http.StatusFound)
-		return
-	}
 }
 
 // json index tournaments handler
