+++ main.go
@@ -22,9 +22,7 @@ import (
 	"github.com/santiaago/purple-wing/helpers/handlers"
 
 	invitectrl "github.com/santiaago/purple-wing/controllers/invite"
-	pagesctrl "github.com/santiaago/purple-wing/controllers/pages"
 	sessionsctrl "github.com/santiaago/purple-wing/controllers/sessions"
-	settingsctrl "github.com/santiaago/purple-wing/controllers/settings"
 	teamrelsctrl "github.com/santiaago/purple-wing/controllers/teamrels"
 	teamsctrl "github.com/santiaago/purple-wing/controllers/teams"
 	tournamentrelsctrl "github.com/santiaago/purple-wing/controllers/tournamentrels"
@@ -36,54 +34,6 @@ import (
 // entry point of application
 func init() {
 	h := new(handlers.RegexpHandler)
-
-	// usual pages
-	// server handler
-	h.HandleFunc("/", pagesctrl.TempHome)
-	h.HandleFunc("/m/?", pagesctrl.Home)
-	h.HandleFunc("/m/about/?", pagesctrl.About)
-	h.HandleFunc("/m/contact/?", pagesctrl.Contact)
-	// session
-	h.HandleFunc("/m/auth/?", sessionsctrl.Authenticate)
-	h.HandleFunc("/m/auth/facebook/?", sessionsctrl.AuthenticateWithFacebook)
-	h.HandleFunc("/m/auth/facebook/callback/?", sessionsctrl.FacebookAuthCallback)
-	h.HandleFunc("/m/auth/google/?", sessionsctrl.AuthenticateWithGoogle)
-	h.HandleFunc("/m/auth/google/callback/?", sessionsctrl.GoogleAuthCallback)
-	h.HandleFunc("/m/auth/twitter/?", sessionsctrl.AuthenticateWithTwitter)
-	h.HandleFunc("/m/auth/twitter/callback/?", sessionsctrl.TwitterAuthCallback)
-	h.HandleFunc("/m/logout/?", handlers.User(sessionsctrl.SessionLogout))
-	// user
-	h.HandleFunc("/m/users/[0-9]+/?", handlers.User(usersctrl.Show))
-	// admin
-	h.HandleFunc("/m/a/?", handlers.Admin(usersctrl.AdminShow))
-	h.HandleFunc("/m/a/users/?", handlers.Admin(usersctrl.AdminUsers))
-	// team
-	h.HandleFunc("/m/teams/?", handlers.User(teamsctrl.Index))
-	h.HandleFunc("/m/teams/new/?", handlers.User(teamsctrl.New))
-	h.HandleFunc("/m/teams/[0-9]+/?", handlers.User(teamsctrl.Show))
-	h.HandleFunc("/m/teams/[0-9]+/edit/?", handlers.User(teamsctrl.Edit))
-	h.HandleFunc("/m/teams/destroy/[0-9]+/?", handlers.User(teamsctrl.Destroy))
-	h.HandleFunc("/m/teams/[0-9]+/invite/?", handlers.User(teamsctrl.Invite))
-	h.HandleFunc("/m/teams/[0-9]+/request/?", handlers.User(teamsctrl.Request))
-	// tournament
-	h.HandleFunc("/m/tournaments/?", handlers.User(tournamentsctrl.Index))
-	h.HandleFunc("/m/tournaments/new/?", handlers.User(tournamentsctrl.New))
-	h.HandleFunc("/m/tournaments/[0-9]+/?", handlers.User(tournamentsctrl.Show))
-	h.HandleFunc("/m/tournaments/[0-9]+/edit/?", handlers.User(tournamentsctrl.Edit))
-	h.HandleFunc("/m/tournaments/destroy/[0-9]+/?", handlers.User(tournamentsctrl.Destroy))
-	// relationships
-	h.HandleFunc("/m/teamrels/create/?", handlers.User(teamrelsctrl.Create))
-	h.HandleFunc("/m/teamrels/destroy/?", handlers.User(teamrelsctrl.Destroy))
-	h.HandleFunc("/m/tournamentrels/create/?", handlers.User(tournamentrelsctrl.Create))
-	h.HandleFunc("/m/tournamentrels/destroy/?", handlers.User(tournamentrelsctrl.Destroy))
-	h.HandleFunc("/m/tournamentteamrels/create/?", handlers.User(tournamentteamrelsctrl.Create))
-	h.HandleFunc("/m/tournamentteamrels/destroy/?", handlers.User(tournamentteamrelsctrl.Destroy))
-	// settings
-	h.HandleFunc("/m/settings/edit-profile/?", handlers.User(settingsctrl.Profile))
-	h.HandleFunc("/m/settings/networks/?", handlers.User(settingsctrl.Networks))
-	h.HandleFunc("/m/settings/email/?", handlers.User(settingsctrl.Email))
-	// invite
-	h.HandleFunc("/m/invite/?", handlers.User(invitectrl.Email))
 
 	// ------------- Json Server -----------------
 
@@ -113,7 +63,6 @@ func init() {
 	h.HandleFunc("/j/tournaments/search/?", handlers.ErrorHandler(handlers.Authorized(tournamentsctrl.SearchJson)))
 	h.HandleFunc("/j/tournaments/candidates/[0-9]+/?", handlers.ErrorHandler(handlers.Authorized(tournamentsctrl.CandidateTeamsJson)))
 	h.HandleFunc("/j/tournaments/[0-9]+/participants/?", handlers.ErrorHandler(handlers.Authorized(tournamentsctrl.ParticipantsJson)))
-
 	// relationships - json
 	h.HandleFunc("/j/teamrels/create/[0-9]+/?", handlers.ErrorHandler(handlers.Authorized(teamrelsctrl.CreateJson)))
 	h.HandleFunc("/j/teamrels/destroy/[0-9]+/?", handlers.ErrorHandler(handlers.Authorized(teamrelsctrl.DestroyJson)))
